fully switch to AllenInstitute ecephys_spike_sorting
