+++ neuropixel_pipeline/modules/atlab/__main__.py
@@ -161,7 +161,6 @@ class AtlabParams(BaseModel):
         logging.info("done with clustering section")
 
         logging.info("starting post-clustering section")
-        # ephys.WaveformSet.populate()
         ephys.QualityMetrics.populate()
         logging.info("done with post-clustering section")
